minecraft-web-manager/backend-python/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-socketio==5.11.0
sqlalchemy==2.0.25
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
python-multipart==0.0.6
PyYAML==6.0.1
python-dotenv==1.0.0
jinja2==3.1.3
psutil==5.9.8
aiofiles==23.2.1
